python/requirements.txt
pyside==5.15.2
numpy==1.20.1
